TBDM-CEUR-WS-main/search-app/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.3
pymongo==4.8.0
python-dotenv==1.0.1
sentence-transformers==3.0.1
google-generativeai==0.7.1